Don't share data sources from local system.
And tell the user they won't be shared.  For example, if the user drags/drops a GeoJSON file, that data source will not be shared because we either need server involvement, or we need to embed the GeoJSON itself in the URL, which is impractical.  So for now we just tell the user that data source will not be shared.