Don't use gigs of memory to load WMS metadata.
When a GetCapabilities has an element with a huge amount of text, that text gets stored in an object for which "typeof value === 'string'" is false but "value instance String" is true.  Since we were only checking for strings with the former test, these huge strings would instead be interpreted as objects and each letter would a separate sub-property.  This caused memory usage to explode (not to mention it made the metadata display pretty useless).  So this change adds the latter test as well.

Also fixed a typo in an error message.