Update ckanext-cesiumpreview for CKAN 2.3
Updates ckanext-cesiumpreview to use JSON config string.

Default view extent is set to the dataset extent recorded in the
"spatial" field in CKAN if available.

WMS resources can use "wms_layer" field to specify a specific layer
to be included in the config and be loaded by default.

CKAN 2.3 includes Resource Views feature which allows adding of
a view including manually to CSV files when the data owner knows the
location columns are appropriate for use in spatial preview.