Add loading indicator to catalog item checkbox
This commit uses css to animate the checkbox associated with data
catalog items. It provides three states - unchecked, loading and
loading. The checkbox itself is created with SVG.

The loading animation animates the dash-offset property to provide the
motion. The checked property uses an animation to draw the stroke path
and then fade in the fill. It's a less-effective animation at smaller
sizes and may (or may not) require tweaking.