Refactor loading indicator to use ViewModes. Now Viewing uses isActive.
CsvVariable items do not have an isLoading property, and adding
the checking logic into the binding syntax turned everything
ugly. Refactored to use a ViewModel method, which is much cleaner.

Now Viewing items, however, do not have an isEnabled property, so
we simply check the isAcvtive property in the binding.