WebMapServiceCatalogItem now gets legend url from first Style tag
If any style tag is present in the Capabilities of a WMS layer with a
LegendURL, this will be used by default for the legend image. If multiple
<Style>s or <LegendURL>s are present, the first one will be used. If none
are present, the legend URL will revert back to a default constructed
url based on the WMS standard, as it currently does now.

Improved determination of legend url type (a or img tag)

In this commit:
 - Created new LegendUrl model class that encapsulates the legend
   url and its type
 - Collapsed CatalogMember "legendUrl" and "legendUrls" properties
   into just "legendUrls", but data can still be read from json with
   a single "legendUrl" string.
 - Added tests for uniform legendUrls property on catalog items.