Allow HTML inside markdown.
Previously, if we saw an HTML tag anywhere in a possibly-markdown string, we would treat it as pure HTML instead of markdown.  But this is incorrect.  HTML is allowed inside markdown.  With this PR, we only disable markdown rendering if the first non-whitespace character is the start of an HTML tag.