[TerriaJS/geoglam-nm#21] Nominally normalise the size of glyphs connected with bar-chart.svg.
The aim of this commit is to have as close as possible to no effect on the redered output, but make the glyph size occupy as close as reasonably possible to the full drawable area.
To simplify and keep pixel values integer summary-concept .btnBack has had the values rounded to the nearest whole pixel value as the original values don't have 5 as a factor.