[#2761] CatalogItem: Remove isShown from _mirroringTerriaClock so that the times are kept in sync even when the layer is not shown.
Better to be in sync more of the time (always would be ideal...but also there are performance considerations to consider). This means that the time between the timeline and the date picker will be kept in sync while the layer is not shown but is connected to the timeline.