feat: hook viewer change events
Hook viewer change events so they are raised after new cesium or leaflet
viewer is created. Changed home camera to getter/setter so it will
properly preserve home camera. Without that, a new viewer will be
created on initialization while the camera view is still not read from
the init file, and it won't be preserved.