Enable `useDefineForClassFields` in tsconfig.json.
This makes TS emit code that is inline with TC39 proposal.
Enabling this options now, makes sure our code is compatible when TS
eventually makes this option the default. Also, this is recommended by
[mobx6](https://mobx.js.org/installation.html#use-spec-compliant-transpilation-for-class-properties)

One gotcha is when a child class re-declares a field. Previously the field
could be left un-initialized but now the field has to be either initialized or
the `declare` keyword must be used so that TS does not emit code for the field.
This commit includes changes to fix this issue in a few trait-classes where we
override a trait declaration with a new one.