Remove `string-replace-webpack-plugin` to fix critical security warning.
The plugin was mainly used for replacing babel generated getters with `superGet` in
mobx 4. But after mobx upgrade this is no longer required.

It is currently only used for upadting the credit string in Ion.js that is shown when using the
default Ion token. I have dropped that in favour of keeping the build script simpler.