Show inline error when feature info panel chart loading fails.
Current behaviour is to pop an error message to the user, but when
the load function being connected to the react life-cycle methods, it
can be called several times. This results in as many pop ups thrown
to the user. We could appropriately handle this if we re-write the
component as a functional component.

However, I think a better solution might be to show the error message
inline instead of popping it to the user.