Run yarn upgrade
Update to the latest packages
so the output of 'yarn audit'
becomes useful.

There is some kind of incompatibility
with @types/react version 18, so restrict
the packages to the latest version 17
for now.

This also discovered something with
the types of lodash-es, so add a type
signature and a default value to
avoid a TS2322 error.

Before:

140 vulnerabilities found - Packages audited: 1918
Severity: 7 Low | 47 Moderate | 70 High | 16 Critical

After:

35 vulnerabilities found - Packages audited: 1976
Severity: 2 Low | 13 Moderate | 16 High | 4 Critical