Upgrade to TS 5.6.
Mainly implements a workaround for error TS2855 that prevents accessing parent class
fields as `super.field` [introduced in TS
5.3](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-3.html#checks-for-super-property-accesses-on-instance-fields). In
valid JS, super can only be used to access get/set accessors or methods from
parent classes that are declared on the prototype, but not instance
fields (declared as `x = "something"`). In our case though, traits are declared
as getters and not fields, however because we use mapped types to mixin the
traits into a class and mapped types does not allow specifying getter/setter
variants, TS considers our traits to be fields and raises TS2855.

The workaround in this PR defines these erroring traits as getters which seems
to convince TS (at least for now) that they are valid accessors eventhough they
still are typed using mapped.

Hopefully in the future, we get this TS feature implementing [variant accessors
for mapped types](https://github.com/microsoft/TypeScript/issues/43826), then
we can get rid of this workaround.