Replace glob-all with fast-glob
I got this warning when running
some command:

npm warn deprecated inflight@1.0.6: This module is not supported, and leaks memory. Do not use it. Check out lru-cache if you want a good and tested way to coalesce async requests by a key value, which is much more comprehensive and powerful.

and inflight is a dependency of glob,
which glob-all depends on. So replace
glob-all with fast-glob that both seems
fast and is already used by some other
dependencies.