Update to ESLint 9.17.0
Update to ESLint 9 since
ESLint 8 reached end-of-life
on 2024-10-05 and is no longer
maintained.

The main part of this PR is
the new flat configuration
file, which is quite close to
a 1:1 migration of the previous
non-flat config. The majority
of lines is the rules section,
which is the same as before
except for the now deprecated
rules that I removed.