eslint: warn about unused directives
Make ESLint warn when directives
are not used, and remove the unused
directives from the source code.

The no-empty rule needs a comment
rather than an ignore directive
in empty catch blocks, so tweak
the config of the rule to admit
empty catch blocks, and report
on other empty bodies.