gulpfile.js: update to V4 syntax
The task() API is not recommended
any more according to the Gulp
documentation, so use the V4 syntax
where a regular function is used
and tasks are the functions that are
exported from the gulpfile.

This commit also hides what seems to
be gulp-internal targets, and adds
help texts for the public targets
that is shown in "gulp --tasks".