jshintrc: set esversion to 11
WebStorm interprets the existence
of .jshintrc as an intention to use
JSHint, and therefore gives warnings
about syntax that is not supported by
the language version used in
the repository.

Set the version to 11 which is
the highest version documented at:
https://jshint.com/docs/options/#esversion
so JSHint stops complaining about
various things that seem to work
in CI.

Setting the esversion gives a complaint
that esnext is incompatible with esversion.
The esnext option is deprecated according to
https://jshint.com/docs/options/#esnext
so remove it.