Replace Prettier with Biome
The Prettier challenge was a reward
of $20 000 for a formatter written
in Rust that passes 95% of the Prettier
test suite. Biome won that:

https://prettier.io/blog/2023/11/27/20k-bounty-was-claimed

This PR does a "yarn biome migrate prettier"
and tweaks the result so the formatting
differences are minimal to Prettier 2.7.1
that is currently used. The performance
of Prettier 3.4.2 is comparable to 2.7.1,
and Biome is quite a bit faster than both
of them:

$ time -p yarn prettier
...
wwwroot/test/Terria/applyInitData/MapServer/mapServerWithError.json 1ms
Done in 19.41s.
real 19.54
user 28.86
sys 2.23

$ time -p yarn format
yarn run v1.22.22
$ biome format --write .
Formatted 1376 files in 251ms. Fixed 70 files.
Done in 0.35s.
real 0.49
user 1.83
sys 0.27