lib/ReactViews: remove unused files
Guidance.jsx and GuidanceDot.jsx seems
to have been unused since at least
commit 62d409cd6c in 2019.

usePrevious.ts entered the tree in
commit 9cbbc8976c3, but no longer seems
to be used.

CatalogShortcut.tsx entered the tree
in commit 287fb3791 in TerriaJS 5.x,
but does not seem to be used, and
it seems ToolButton.tsx has a similar
history.

SplitPoint.jsx entered the tree in
commit 40bd098b29, and also seems to
be unused.