gulpfile: import plugin-error first
Hoist the multiple imports of plugin-error
to a single top level import. Also remove
the "showStack: false" configuration to each
call since that has been the default
for ~10 years, break up the long string
to fit in a line, and remove the
questionmark at the end of the error
message.