Track initial camera view for viewers and return it if the camera hasn't changed.
Previously, the camera view would drift when switching between viewer
modes. This is because each time we switch the viewer, we get the current view
from leaflet/cesium scene which renders the map differently. This can be a
problem when restoring share links and the viewer is first initialized in
3d/cesium mode and then switches to leaflet/2d mode.

This change makes sure that the view remains stable if the camera hasn't
changed after setting the initial view.