Fix failing specs.
The specs fail for some reason in Firefox 145>=. The issue seems to be that the
last argument of `CreateModel` constructor - the strata param - is being passed
in as `this` instead of 'undefined` resulting in `this` getting set as the
`strata`. Using `varargs` seems to fix it.

Not clear why it happens but it does this only in our test environment, the app
itself appears to work fine in Firefox 145 without the fix.